Letest from inid master

+++ 3rdparty/indi-eqmod/simulator/skywatcher-simulator.cpp
@@ -4,6 +4,7 @@
 #include <indidevapi.h>
 
 #include <string.h>
+#include <stdint.h>
 
 void SkywatcherSimulator::send_byte(unsigned char c)
 {
@@ -129,7 +130,7 @@ void SkywatcherSimulator::setupDE(unsigned int nb_teeth, unsigned int gear_ratio
 
 void SkywatcherSimulator::compute_timer_ra(unsigned int wormperiod)
 {
-    unsigned long n = (wormperiod * MUL_RA);
+    uint32_t n = (wormperiod * MUL_RA);
     n += ((wormperiod * REM_RA) / ra_steps_worm);
     ra_period = n;
     //IDLog("Simulator RA Worm period = %d, Microstep timer period = %d µs\n", wormperiod, ra_period);
@@ -137,7 +138,7 @@ void SkywatcherSimulator::compute_timer_ra(unsigned int wormperiod)
 
 void SkywatcherSimulator::compute_timer_de(unsigned int wormperiod)
 {
-    unsigned long n = (wormperiod * MUL_DE);
+    uint32_t n = (wormperiod * MUL_DE);
     n += ((wormperiod * REM_DE) / de_steps_worm);
     de_period = n;
     //IDLog("Simulator DE Worm period = %d, Microstep timer period = %d µs\n", wormperiod, de_period);
